wip threebot extend enhancments
