+++ jumpscale/packages/threebot_deployer/chats/extend.py
@@ -1,10 +1,9 @@
 from jumpscale.loader import j
-from jumpscale.sals.chatflows.chatflows import StopChatFlow, chatflow_step
-from jumpscale.sals.marketplace.chatflow import MarketPlaceChatflow
-from jumpscale.sals.marketplace import deployer, solutions
+from jumpscale.sals.chatflows.chatflows import chatflow_step
+from jumpscale.sals.marketplace import MarketPlaceAppsChatflow, solutions
 
 
-class ExtendThreebot(MarketPlaceChatflow):
+class ExtendThreebot(MarketPlaceAppsChatflow):
 
     title = "Extend 3Bot"
     steps = ["select_threebot", "new_expiration", "solution_extension", "success"]
@@ -12,7 +11,7 @@ class ExtendThreebot(MarketPlaceChatflow):
     @chatflow_step(title="Select 3Bot")
     def select_threebot(self):
         self._validate_user()
-        self.query = {"cru": 2, "mru": 2, "sru": 2}
+        self.query = {"cru": 1, "mru": 1, "sru": 2}
         self.threebot_name = self.user_info()["username"]
         self.explorer = j.core.identity.me.explorer
         self.threebot_workloads = solutions.list_solutions(self.threebot_name, "threebot")
@@ -23,32 +22,6 @@ class ExtendThreebot(MarketPlaceChatflow):
         )
         self.pool_id = self._threebots_dict[self.threebot_selected]["Pool"]
         self.pool = j.sals.zos.pools.get(self.pool_id)
-
-    @chatflow_step(title="New Expiration")
-    def new_expiration(self):
-        DURATION_MAX = 9223372036854775807
-        if self.pool.empty_at < DURATION_MAX:
-            # Pool currently being consumed (compute or storage), default is current pool empty at + 65 mins
-            min_timestamp_fromnow = self.pool.empty_at - j.data.time.get().timestamp
-            default_time = self.pool.empty_at + 3900
-        else:
-            # Pool not being consumed (compute or storage), default is in 14 days (60*60*24*14 = 1209600)
-            min_timestamp_fromnow = None
-            default_time = j.data.time.get().timestamp + 1209600
-        self.expiration = deployer.ask_expiration(self, default_time, min=min_timestamp_fromnow)
-
-    @chatflow_step(title="Payment")
-    def solution_extension(self):
-        self.currencies = ["TFT"]
-        self.pool_info, self.qr_code = deployer.extend_solution_pool(
-            self, self.pool_id, self.expiration, self.currencies, **self.query
-        )
-
-        result = deployer.wait_pool_payment(
-            self, self.pool_id, qr_code=self.qr_code, trigger_cus=self.pool.cus + 1, trigger_sus=self.pool.sus + 1
-        )
-        if not result:
-            raise StopChatFlow(f"Waiting for pool payment timedout. pool_id: {self.pool_id}")
 
     @chatflow_step(title="Success", final_step=True)
     def success(self):
